Refactor populateDist
