+++ packages/cursorless-vscode/src/scripts/populateDist/context.ts
@@ -1,0 +1,13 @@
+export interface Context {
+  /** Whether we're deploying the final extension */
+  isDeploy: boolean;
+
+  /**
+   * If `true`, then we override the extension id in order to install the
+   * extension locally without new Cursorless version releases clobbering it.
+   */
+  isForLocalInstall: boolean;
+
+  /** Whether we're running in CI */
+  isCi: boolean;
+}
