+++ packages/cursorless-vscode/src/scripts/populateDist/index.ts
@@ -1,0 +1,3 @@
+import { run } from "./populateDist";
+
+run();
