+++ packages/cursorless-vscode/src/scripts/populateDist/runCommand.ts
@@ -1,0 +1,14 @@
+import { exec } from "child_process";
+import { promisify } from "util";
+
+const execAsync = promisify(exec);
+
+export async function runCommand(command: string) {
+  const { stdout, stderr } = await execAsync(command);
+
+  if (stderr) {
+    throw new Error(stderr);
+  }
+
+  return stdout;
+}
